pyproject.toml
[tool.poetry]
name = "lido-dao-scripts"
version = "0.1.0"
description = "Repository for Lido DAO Aragon voting scripts."
authors = ["Lido devteam"]
license = "MIT"
readme = "README.md"
homepage = "https://mainnet.lido.fi"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-brownie = "1.20.2"
avotes-parser-core = "~0.5.6"
hexbytes = "0.3.1"
ipfs-cid = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "~24.2.0"
pre-commit = "~2.19.0"
pytest-check = "^2.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:abi.(decode|encode):DeprecationWarning",
    "ignore:rpc.(snapshot|revert):FutureWarning",
]
